don't try to download missing counties
If an address was geocoded to the centroid of a zip code or city, it is possible that the centroid would actually not fall within a valid county boundary and so the code would fail when attempting to download the tract shapefiles for a county with a FIPS of
`NA`. This change omits `NA` from the counties needed to download.  The census tract will then remain as `NA` in the output.